requirements.txt
# PostgreSQL Query Analyzer Bot - Python Dependencies

# Slack SDK and Bolt Framework
slack-bolt==1.18.0
slack-sdk==3.24.0

# Async HTTP Client
aiohttp==3.9.1

# Environment Variables
python-dotenv==1.0.0

# Optional: For enhanced logging
colorlog==6.7.0

# Optional: For metrics and monitoring
# prometheus-client==0.19.0

# Development dependencies (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# flake8==6.1.0
# mypy==1.7.0